extra work to extend functionality
Added some extra functionality with a search bar.  A user can type in characters and results will show immediately or they can paste in with their mouse to find results.  If no results are found a message will appear letting the user know and the pagination changes based on how many results are returned from the search.